Cleanup CrystalEdit: `size_t` related issues
 * eliminate conversion warnings for `size_t` related expressions.
 * convert various `int` and `long` to `size_t`
 * add `static_cast<int>` or add `static_cast<long>` as appropriate
 * classical `(int)` casting is used in ASSERT, for simple convenience.
 * add a few ASSERT() statements to assure that various `size_t` values
	remain within the positive `int` value range.
 * Since `-1` is often used as a "size flag", these changes now
	explicitly test for `x != -1` rather than `x >= 0`

While this commit changes various parameters and data items to `size_t`,
the focus of these changes was to completely eliminate the various
warnings. This code is still fundamentally restricted to 2GB of text.

There is still much work to do to generalize the fundamentals of
CrystalEdit to be fully `size_t` capable (for the X64 platform).  Two
items of important note:
 	1. MFC's CPOINT structure is limited to `LONG` values.
	   The `.x` member should become `size_t`.
	2. MFC's CSTRING data is limited to `int` maximum length