Cleanup: Misc `size_t` related issues
 * eliminate conversion warnings for `size_t` related expressions.
 * convert various `int` and `long` to `size_t`
 * add `static_cast<int>` or add `static_cast<long>` as appropriate
 * classical `(int)`-style casting is used in *.c files, of course

While this commit changes various parameters and data items to `size_t`,
the focus of these changes was to completely eliminate the various
warnings. This code is still fundamentally restricted to 2GB of text.

There is still much work to do to generalize the fundamentals of
WinMerge to be fully `size_t` capable (for the X64 platform).