Cleanup2: Uninitialized Local Variables
I've set warning level = 4 for the WinMerge project (but not for the
three 'poco' projects).  It generates LOTS of warnings; many categories
are not interesting and I certainly do not recommend that the Level=4
become an official setting.  But some warning categories do seem
important, either functionally or aesthetically.   So, I'll slowly be
working on the ones that I think are interesting...

 * L4 warning C4701: "potentially uninitialized local variable used".
This commit eliminates all occurrences from WinMerge and UnitTests
projects (but not from the 'poco' projects).